requirements.txt
fastapi
uvicorn
langchain
chromadb
transformers
sentence-transformers
unstructured
pdfminer.six
python-docx
